add loading animation for switch app
