+++ main/public/index.html
@@ -25,10 +25,13 @@
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
     <title>deri-fi</title>
+    <!-- <link type="text/css" rel="stylesheet" href="%PUBLIC_URL%/index.css" /> -->
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
+    <div id="root">
+      <!-- <img class='loading-banner' src="%PUBLIC_URL%/loading.png"/> -->
+    </div>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
