+++ main/src/components/Loading/Loading.js
@@ -1,7 +1,10 @@
 import './loading.scss'
+import loadingPic from '../../assets/img/loading.png'
 
 export default function Loading(){
   return (
-    <div className='loading'></div>
+    <div className='loading-pic'>
+      <img src={loadingPic} />
+    </div>
   )
 }
