+++ main/src/components/Loading/loading.scss
@@ -1,0 +1,11 @@
+.loading-pic {
+    display: flex;
+    justify-content: center;
+    top: calc( 50% - 150px);
+    left: calc( 50% - 150px);
+    transform: translate(-50%, -50%);
+    transform: translate(-50%, -50%);
+    position: absolute;
+    top: 50%;
+    left: 50%;
+}
